Add tar_add_data() function.
The tar_add_data function directly adds data to a tar file.
(Normally a file is required to add data to the tar file.)

The adding of the LICENSE file is updated to use this new method,
even though currently the LICENSE is stored in a file, this will
change in the updated version.