Ensure prj-test also correctly returns an error value to the shell.
Rather than duplicating the code in x_test, in prj_test, the underlying
run_module_tests_with_args/run_module_tests function is changed to
directly return a value suitable for returning the the shell.

Rather than simply returning 0 on success, 1 on failure, the number
of failed tests is returned to the shell.