Ensure x tests work correctly on TEAMCITY.
On TEAMCITY there is no 'git' repo when the tests are run, so
avoid running git related unit tests on TeamCity.

This isn't ideal; ideally TeamCity would have a way to access
git, however this isn't critical, and x.py is only an internal
tool, so if the git support breaks it will not easily go unnoticed.