Simplify parser/renderengine internals.
Previously there was an explicit distinction between a literal
and escaped interpolation. With the addition of formatters generally
for interpolation there is the opportunity to consolidate this
internally.

Interally, literals are now handled as formatters with the key
'literal'. In this manner EscapeNode and LiteralNode can be merged
to a single InterpolateNode, reducing duplication internally, and
simplifying internal interfaces.