Ensure preemption is enabled on first switch-to a task.
Previously it was possible that the first switch to a task
would leave preemption disable.

This is because the only preempt enable path is on the exit
from yield(), however when first switching to to a task, it
is switched directly to the entry function, bypassing the yield
code.

This introduces a simple trampoline for each task that ensures
preempt_enable is called before calling the task's main function.