Wrap task functions in a trampoline.
The trampoline waits on the start signal (where appropriate) before
running the task function.

It is possible that auto-start tasks don't need a trampoline, however
to have enhanced error handling a trampoline can perform a fatal
error (in a future task) if the task function ever returns, so a
trampoline is created for all tasks.