Add support for constraints in the schema definition.
Although element specific constraints (such as the type
and optional tag) provide a reasonable level of configuration
checking, there are times when inter-element constraints are
required. There more general constraints concept supports this.

Currently the only supported type of contstraint is the 'one-of'
constraint, which indicates that exactly one-of the listed elements
must be specified in the configuration.

It is possible that a less powerful 'choice' element could have
been added instead, however constraints provided the possibility
of a more general mechanism that can be extended in the future.

Additionally, adding the constraint type required less rework of
existing code than adding a choice element.